refactor: Move uninstall package helpers
Moves `get_install_ids` onto `Manifest` in parallel with
`Manifest.install`.

The errors `PackageNotFound` and `MultiplePackagesMatch` have been moved
further down into `ManifestError` to match. These are distinct from
`PkgOrGroupNotFound` because we only handle package names for
`uninstall` and `get_install_ids`, unlike `upgrade`. I've changed the
user facing message in order to match `upgrade` and because we don't
know that the operation is `uninstall` this far down.

Moves `get_includes_for_packages` and `get_include_for_package` onto
`Compose` in parallel with `Compose.include`.

The error `PackageOnlyIncluded` has been grouped in `UninstallError`
which is then wrapped by `CoreEnvironmentError` in the same way as
`UpgradeError`.

The wrapping of error messages still feels complicated but I'm not sure
there's much more to do prior ticket 1965.
