+++ cli/flox/src/utils/errors.rs
@@ -277,9 +277,6 @@ pub fn format_core_error(err: &CoreEnvironmentError) -> String {
 
             Please ensure that you have write permissions to '.flox/env/manifest.toml'.
         "},
-        CoreEnvironmentError::PackageNotFound(_) => display_chain(err),
-        CoreEnvironmentError::PackageOnlyIncluded(_, _) => display_chain(err),
-        CoreEnvironmentError::MultiplePackagesMatch(_, _) => display_chain(err),
 
         // internal error, a bug if this happens to users!
         CoreEnvironmentError::BadLockfilePath(_) => display_chain(err),
@@ -304,6 +301,7 @@ pub fn format_core_error(err: &CoreEnvironmentError) -> String {
                     $ flox upgrade
             "},
         },
+        CoreEnvironmentError::UninstallError(_) => display_chain(err),
         // User facing
         CoreEnvironmentError::Services(err) => display_chain(err),
